find_timeline/requirements.txt
keras==2.11.0
numpy==1.22.3
pandas==1.5.2
requests==2.28.1
scikit_learn==1.2.0
tensorflow_macos==2.11.0
torch==1.12.1
tqdm==4.64.1
trafilatura==1.4.0
transformers==4.24.0
